refactor(initialization): split genesis phase
